avoid NullPointerException
git-svn-id: https://server.mycore.de/svn/mycore/trunk@12565 7089d063-bc14-496d-ac90-53de1872eb6c

+++ modules/module-indexing/sources/org/mycore/frontend/indexbrowser/MCRGoogleSitemapCommon.java
@@ -260,11 +260,13 @@ public final class MCRGoogleSitemapCommon {
         }
         File dir = new File(sb.toString());
         File[] li = dir.listFiles();
-        for (File fi : li) {
-            if (fi.getName().startsWith("sitemap_google")) {
-                LOGGER.debug("Remove file " + fi.getName());
-                fi.delete();
-            }
+        if(li!=null){
+        	for (File fi : li) {
+        		if (fi.getName().startsWith("sitemap_google")) {
+        			LOGGER.debug("Remove file " + fi.getName());
+        			fi.delete();
+        		}
+        	}
         }
     }
 }
